requirements.txt
black==22.1
click==8.0
flake8==4.0
flask==2.0
jinja2==3.0
mypy==0.940
pillow==9.0
pre-commit==2.17
requests==2.27
